# helpfirst-be/db/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.25.0

package db

import (
	"database/sql"
	"time"

	"github.com/google/uuid"
)

type Reports struct {
	ID        uuid.UUID `json:"id"`
	Uid       uuid.UUID `json:"uid"`
	Title     string    `json:"title"`
	Type      string    `json:"type"`
	Level     string    `json:"level"`
	Address   string    `json:"address"`
	Lat       float64   `json:"lat"`
	Lng       float64   `json:"lng"`
	CreatedAt time.Time `json:"createdAt"`
	UpdatedAt time.Time `json:"updatedAt"`
}

type Users struct {
	Uid               uuid.UUID      `json:"uid"`
	Name              sql.NullString `json:"name"`
	Email             string         `json:"email"`
	Avatar            sql.NullString `json:"avatar"`
	Password          sql.NullString `json:"password"`
	Provider          string         `json:"provider"`
	CreatedAt         time.Time      `json:"createdAt"`
	UpdatedAt         time.Time      `json:"updatedAt"`
	PasswordChangedAt time.Time      `json:"passwordChangedAt"`
}